WIP: test RU against SH
