+++ heudiconv_app/bids-jsons/site-RU_T1w.json
@@ -1,0 +1,87 @@
+{
+    "AcquisitionDuration": 346.597,
+    "AcquisitionMatrixPE": 256,
+    "AcquisitionNumber": 1,
+    "AcquisitionTime": "11:53:15.110000",
+    "BaseResolution": 256,
+    "BodyPartExamined": "BRAIN",
+    "CoilCombinationMethod": "Adaptive Combine",
+    "CoilString": "HeadNeck_64",
+    "ConversionSoftware": "dcm2niix",
+    "ConversionSoftwareVersion": "v1.0.20220720",
+    "DeviceSerialNumber": "66022",
+    "DwellTime": 1.22e-05,
+    "EchoTime": 0.00307,
+    "FlipAngle": 8,
+    "ImageOrientationPatientDICOM": [
+        1,
+        0,
+        0,
+        0,
+        1,
+        0
+    ],
+    "ImageType": [
+        "ORIGINAL",
+        "PRIMARY",
+        "M",
+        "NONE"
+    ],
+    "ImageTypeText": [
+        "ORIGINAL",
+        "PRIMARY",
+        "M",
+        "ND",
+        "NORM"
+    ],
+    "ImagingFrequency": 123.257,
+    "InPlanePhaseEncodingDirectionDICOM": "COL",
+    "InstitutionAddress": "East Beltline NE 2750,Grand Rapids,MI,US,49503",
+    "InstitutionName": "Spectrum ICCB Campus",
+    "InversionTime": 1.06,
+    "MRAcquisitionType": "3D",
+    "MTState": false,
+    "MagneticFieldStrength": 3,
+    "Manufacturer": "Siemens",
+    "ManufacturersModelName": "MAGNETOM Prisma",
+    "MatrixCoilMode": "GRAPPA",
+    "Modality": "MR",
+    "NonlinearGradientCorrection": false,
+    "ParallelAcquisitionTechnique": "GRAPPA",
+    "ParallelReductionFactorInPlane": 2,
+    "PartialFourier": 1,
+    "PartialFourierDirection": "SLICE_SELECT",
+    "PatientPosition": "HFS",
+    "PercentPhaseFOV": 100,
+    "PercentSampling": 100,
+    "PhaseResolution": 1,
+    "PixelBandwidth": 160,
+    "ProcedureStepDescription": "Physicist ABCD",
+    "ProtocolName": "T1_MPRAGE",
+    "PulseSequenceDetails": "%SiemensSeq%\\tfl",
+    "PulseSequenceName": "*tfl3d1_16",
+    "ReceiveCoilActiveElements": "HC1-7",
+    "ReceiveCoilName": "HeadNeck_64",
+    "ReconMatrixPE": 256,
+    "RefLinesPE": 24,
+    "RepetitionTime": 2.5,
+    "ScanOptions": "IR",
+    "ScanningSequence": "GR\\IR",
+    "SequenceVariant": "SK\\SP\\MP",
+    "SeriesDescription": "T1_MPRAGE_ND",
+    "SeriesNumber": 2,
+    "ShimSetting": [
+        -1200,
+        -11047,
+        2840,
+        152,
+        177,
+        -293,
+        146,
+        43
+    ],
+    "SliceThickness": 1,
+    "SoftwareVersions": "syngo MR XA30",
+    "StationName": "AWP66022",
+    "TxRefAmp": 147.376
+}
