+++ heudiconv_app/bids-jsons/site-RU_acq-dwib0_dir-AP_epi.json
@@ -1,0 +1,178 @@
+{
+    "AcquisitionDuration": 37.127,
+    "AcquisitionMatrixPE": 140,
+    "AcquisitionNumber": 1,
+    "AcquisitionTime": "11:59:54.720000",
+    "BandwidthPerPixelPhaseEncode": 10.352,
+    "BaseResolution": 140,
+    "BodyPartExamined": "BRAIN",
+    "CoilCombinationMethod": "Adaptive Combine",
+    "CoilString": "HeadNeck_64",
+    "ConversionSoftware": "dcm2niix",
+    "ConversionSoftwareVersion": "v1.0.20220720",
+    "DerivedVendorReportedEchoSpacing": 0.000689998,
+    "DeviceSerialNumber": "66022",
+    "DiffusionScheme": "Monopolar",
+    "DwellTime": 2.1e-06,
+    "EchoTime": 0.088,
+    "EchoTrainLength": 105,
+    "EffectiveEchoSpacing": 0.000689998,
+    "FlipAngle": 90,
+    "ImageOrientationPatientDICOM": [
+        1,
+        0,
+        0,
+        0,
+        1,
+        0
+    ],
+    "ImageType": [
+        "ORIGINAL",
+        "PRIMARY",
+        "DIFFUSION",
+        "NONE"
+    ],
+    "ImageTypeText": [
+        "ORIGINAL",
+        "PRIMARY",
+        "DIFFUSION",
+        "NONE",
+        "ND"
+    ],
+    "ImagingFrequency": 123.257,
+    "InPlanePhaseEncodingDirectionDICOM": "COL",
+    "InstitutionAddress": "East Beltline NE 2750,Grand Rapids,MI,US,49503",
+    "InstitutionName": "Spectrum ICCB Campus",
+    "IntendedFor": [
+        "ses-221108/dwi/sub-sh_ses-221108_dwi.nii.gz"
+    ],
+    "MRAcquisitionType": "2D",
+    "MTState": false,
+    "MagneticFieldStrength": 3,
+    "Manufacturer": "Siemens",
+    "ManufacturersModelName": "MAGNETOM Prisma",
+    "MatrixCoilMode": "SENSE",
+    "Modality": "MR",
+    "NonlinearGradientCorrection": false,
+    "PartialFourier": 0.75,
+    "PartialFourierDirection": "PHASE",
+    "PatientPosition": "HFS",
+    "PercentPhaseFOV": 100,
+    "PercentSampling": 100,
+    "PhaseEncodingDirection": "j-",
+    "PhaseResolution": 1,
+    "PixelBandwidth": 1701,
+    "ProcedureStepDescription": "Physicist ABCD",
+    "ProtocolName": "DWI_B0_AP",
+    "PulseSequenceDetails": "%SiemensSeq%\\ep2d_diff",
+    "PulseSequenceName": "*epse2d1_140",
+    "ReceiveCoilActiveElements": "HC1-7",
+    "ReceiveCoilName": "HeadNeck_64",
+    "ReconMatrixPE": 140,
+    "RepetitionTime": 12.3,
+    "ScanOptions": "PFP\\FS",
+    "ScanningSequence": "EP",
+    "SequenceVariant": "SK\\SP",
+    "SeriesDescription": "DWI_B0_AP",
+    "SeriesNumber": 4,
+    "ShimSetting": [
+        -1205,
+        -11038,
+        2823,
+        143,
+        166,
+        -465,
+        108,
+        64
+    ],
+    "SliceThickness": 1.7,
+    "SliceTiming": [
+        0,
+        6.215,
+        0.1525,
+        6.3675,
+        0.3025,
+        6.5175,
+        0.455,
+        6.67,
+        0.605,
+        6.8225,
+        0.7575,
+        6.9725,
+        0.91,
+        7.125,
+        1.06,
+        7.2775,
+        1.2125,
+        7.4275,
+        1.365,
+        7.58,
+        1.515,
+        7.7325,
+        1.6675,
+        7.8825,
+        1.82,
+        8.035,
+        1.97,
+        8.185,
+        2.1225,
+        8.3375,
+        2.275,
+        8.49,
+        2.425,
+        8.64,
+        2.5775,
+        8.7925,
+        2.7275,
+        8.945,
+        2.88,
+        9.095,
+        3.0325,
+        9.2475,
+        3.1825,
+        9.4,
+        3.335,
+        9.55,
+        3.4875,
+        9.7025,
+        3.6375,
+        9.855,
+        3.79,
+        10.005,
+        3.9425,
+        10.1575,
+        4.0925,
+        10.3075,
+        4.245,
+        10.46,
+        4.395,
+        10.6125,
+        4.5475,
+        10.7625,
+        4.7,
+        10.915,
+        4.85,
+        11.0675,
+        5.0025,
+        11.2175,
+        5.155,
+        11.37,
+        5.305,
+        11.5225,
+        5.4575,
+        11.6725,
+        5.61,
+        11.825,
+        5.76,
+        11.975,
+        5.9125,
+        12.1275,
+        6.065
+    ],
+    "SoftwareVersions": "syngo MR XA30",
+    "SpacingBetweenSlices": 1.7,
+    "StationName": "AWP66022",
+    "TotalReadoutTime": 0.0959097,
+    "TxRefAmp": 147.376,
+    "VendorReportedEchoSpacing": 0.00069
+}
