+++ heudiconv_app/bids-jsons/site-RU_acq-fmrib0_dir-AP_epi.json
@@ -1,0 +1,156 @@
+{
+    "AcquisitionDuration": 20.962,
+    "AcquisitionMatrixPE": 90,
+    "AcquisitionNumber": 1,
+    "AcquisitionTime": "12:13:42.525000",
+    "BandwidthPerPixelPhaseEncode": 21.786,
+    "BaseResolution": 90,
+    "BodyPartExamined": "BRAIN",
+    "CoilCombinationMethod": "Adaptive Combine",
+    "CoilString": "HeadNeck_64",
+    "ConversionSoftware": "dcm2niix",
+    "ConversionSoftwareVersion": "v1.0.20220720",
+    "DerivedVendorReportedEchoSpacing": 0.000510012,
+    "DeviceSerialNumber": "66022",
+    "DiffusionScheme": "Monopolar",
+    "DwellTime": 2e-06,
+    "EchoTime": 0.08,
+    "EchoTrainLength": 90,
+    "EffectiveEchoSpacing": 0.000510012,
+    "FlipAngle": 90,
+    "ImageOrientationPatientDICOM": [
+        1,
+        0,
+        0,
+        0,
+        1,
+        0
+    ],
+    "ImageType": [
+        "ORIGINAL",
+        "PRIMARY",
+        "DIFFUSION",
+        "NONE"
+    ],
+    "ImageTypeText": [
+        "ORIGINAL",
+        "PRIMARY",
+        "DIFFUSION",
+        "NONE",
+        "ND"
+    ],
+    "ImagingFrequency": 123.257,
+    "InPlanePhaseEncodingDirectionDICOM": "COL",
+    "InstitutionAddress": "East Beltline NE 2750,Grand Rapids,MI,US,49503",
+    "InstitutionName": "Spectrum ICCB Campus",
+    "IntendedFor": [
+        "ses-221108/func/sub-sh_ses-221108_task-rest_run-01_bold.nii.gz"
+    ],
+    "MRAcquisitionType": "2D",
+    "MTState": false,
+    "MagneticFieldStrength": 3,
+    "Manufacturer": "Siemens",
+    "ManufacturersModelName": "MAGNETOM Prisma",
+    "MatrixCoilMode": "SENSE",
+    "Modality": "MR",
+    "NonlinearGradientCorrection": false,
+    "PartialFourier": 1,
+    "PatientPosition": "HFS",
+    "PercentPhaseFOV": 100,
+    "PercentSampling": 100,
+    "PhaseEncodingDirection": "j-",
+    "PhaseResolution": 1,
+    "PixelBandwidth": 2778,
+    "ProcedureStepDescription": "Physicist ABCD",
+    "ProtocolName": "fMRI_B0_AP",
+    "PulseSequenceDetails": "%SiemensSeq%\\ep2d_diff",
+    "PulseSequenceName": "*epse2d1_90",
+    "ReceiveCoilActiveElements": "HC1-7",
+    "ReceiveCoilName": "HeadNeck_64",
+    "ReconMatrixPE": 90,
+    "RepetitionTime": 7.1,
+    "ScanOptions": "FS",
+    "ScanningSequence": "EP",
+    "SequenceVariant": "SK\\SP",
+    "SeriesDescription": "fMRI_B0_AP",
+    "SeriesNumber": 13,
+    "ShimSetting": [
+        -1205,
+        -11038,
+        2823,
+        141,
+        166,
+        -466,
+        108,
+        64
+    ],
+    "SliceThickness": 2.4,
+    "SliceTiming": [
+        3.54,
+        0,
+        3.6575,
+        0.1175,
+        3.7775,
+        0.2375,
+        3.895,
+        0.355,
+        4.0125,
+        0.4725,
+        4.13,
+        0.59,
+        4.2475,
+        0.7075,
+        4.3675,
+        0.8275,
+        4.485,
+        0.945,
+        4.6025,
+        1.0625,
+        4.72,
+        1.18,
+        4.8375,
+        1.2975,
+        4.955,
+        1.4175,
+        5.075,
+        1.535,
+        5.1925,
+        1.6525,
+        5.31,
+        1.77,
+        5.4275,
+        1.8875,
+        5.545,
+        2.0075,
+        5.665,
+        2.125,
+        5.7825,
+        2.2425,
+        5.9,
+        2.36,
+        6.0175,
+        2.4775,
+        6.135,
+        2.5975,
+        6.255,
+        2.715,
+        6.3725,
+        2.8325,
+        6.49,
+        2.95,
+        6.6075,
+        3.0675,
+        6.725,
+        3.1875,
+        6.845,
+        3.305,
+        6.9625,
+        3.4225
+    ],
+    "SoftwareVersions": "syngo MR XA30",
+    "SpacingBetweenSlices": 2.4,
+    "StationName": "AWP66022",
+    "TotalReadoutTime": 0.045391,
+    "TxRefAmp": 147.376,
+    "VendorReportedEchoSpacing": 0.00051
+}
