+++ heudiconv_app/bids-jsons/site-RU_acq-fmrib0_dir-PA_epi.json
@@ -1,0 +1,156 @@
+{
+    "AcquisitionDuration": 20.96,
+    "AcquisitionMatrixPE": 90,
+    "AcquisitionNumber": 1,
+    "AcquisitionTime": "12:14:12.502500",
+    "BandwidthPerPixelPhaseEncode": 21.786,
+    "BaseResolution": 90,
+    "BodyPartExamined": "BRAIN",
+    "CoilCombinationMethod": "Adaptive Combine",
+    "CoilString": "HeadNeck_64",
+    "ConversionSoftware": "dcm2niix",
+    "ConversionSoftwareVersion": "v1.0.20220720",
+    "DerivedVendorReportedEchoSpacing": 0.000510012,
+    "DeviceSerialNumber": "66022",
+    "DiffusionScheme": "Monopolar",
+    "DwellTime": 2e-06,
+    "EchoTime": 0.07,
+    "EchoTrainLength": 90,
+    "EffectiveEchoSpacing": 0.000510012,
+    "FlipAngle": 90,
+    "ImageOrientationPatientDICOM": [
+        1,
+        0,
+        0,
+        0,
+        1,
+        0
+    ],
+    "ImageType": [
+        "ORIGINAL",
+        "PRIMARY",
+        "DIFFUSION",
+        "NONE"
+    ],
+    "ImageTypeText": [
+        "ORIGINAL",
+        "PRIMARY",
+        "DIFFUSION",
+        "NONE",
+        "ND"
+    ],
+    "ImagingFrequency": 123.257,
+    "InPlanePhaseEncodingDirectionDICOM": "COL",
+    "InstitutionAddress": "East Beltline NE 2750,Grand Rapids,MI,US,49503",
+    "InstitutionName": "Spectrum ICCB Campus",
+    "IntendedFor": [
+        "ses-221108/func/sub-sh_ses-221108_task-rest_run-01_bold.nii.gz"
+    ],
+    "MRAcquisitionType": "2D",
+    "MTState": false,
+    "MagneticFieldStrength": 3,
+    "Manufacturer": "Siemens",
+    "ManufacturersModelName": "MAGNETOM Prisma",
+    "MatrixCoilMode": "SENSE",
+    "Modality": "MR",
+    "NonlinearGradientCorrection": false,
+    "PartialFourier": 1,
+    "PatientPosition": "HFS",
+    "PercentPhaseFOV": 100,
+    "PercentSampling": 100,
+    "PhaseEncodingDirection": "j",
+    "PhaseResolution": 1,
+    "PixelBandwidth": 2778,
+    "ProcedureStepDescription": "Physicist ABCD",
+    "ProtocolName": "fMRI_B0_PA",
+    "PulseSequenceDetails": "%SiemensSeq%\\ep2d_diff",
+    "PulseSequenceName": "*epse2d1_90",
+    "ReceiveCoilActiveElements": "HC1-7",
+    "ReceiveCoilName": "HeadNeck_64",
+    "ReconMatrixPE": 90,
+    "RepetitionTime": 7.1,
+    "ScanOptions": "FS",
+    "ScanningSequence": "EP",
+    "SequenceVariant": "SK\\SP",
+    "SeriesDescription": "fMRI_B0_PA",
+    "SeriesNumber": 14,
+    "ShimSetting": [
+        -1205,
+        -11038,
+        2823,
+        141,
+        166,
+        -466,
+        108,
+        64
+    ],
+    "SliceThickness": 2.4,
+    "SliceTiming": [
+        3.54,
+        0,
+        3.6575,
+        0.1175,
+        3.775,
+        0.235,
+        3.8925,
+        0.3525,
+        4.0125,
+        0.4725,
+        4.13,
+        0.59,
+        4.2475,
+        0.7075,
+        4.365,
+        0.825,
+        4.4825,
+        0.9425,
+        4.6025,
+        1.0625,
+        4.72,
+        1.18,
+        4.8375,
+        1.2975,
+        4.955,
+        1.415,
+        5.0725,
+        1.5325,
+        5.19,
+        1.6525,
+        5.31,
+        1.77,
+        5.4275,
+        1.8875,
+        5.545,
+        2.005,
+        5.6625,
+        2.1225,
+        5.78,
+        2.2425,
+        5.9,
+        2.36,
+        6.0175,
+        2.4775,
+        6.135,
+        2.595,
+        6.2525,
+        2.7125,
+        6.37,
+        2.8325,
+        6.49,
+        2.95,
+        6.6075,
+        3.0675,
+        6.725,
+        3.185,
+        6.8425,
+        3.3025,
+        6.96,
+        3.4225
+    ],
+    "SoftwareVersions": "syngo MR XA30",
+    "SpacingBetweenSlices": 2.4,
+    "StationName": "AWP66022",
+    "TotalReadoutTime": 0.045391,
+    "TxRefAmp": 147.376,
+    "VendorReportedEchoSpacing": 0.00051
+}
