+++ heudiconv_app/bids-jsons/site-RU_dwi.bval
@@ -1,0 +1,1 @@
+0 0 3000 3000 2000 3000 1000 3000 3000 2000 3000 1000 3000 3000 500 3000 2000 3000 3000 1000 3000 500 3000 3000 2000 3000 1000 3000 3000 2000 3000 1000 3000 3000 0 3000 2000 3000 1000 3000 3000 500 3000 2000 3000 3000 1000 3000 0 3000 3000 2000 3000 1000 3000 3000 2000 3000 1000 3000 3000 500 3000 2000 3000 3000 1000 3000 0 3000 2000 3000 3000 1000 3000 500 3000 3000 2000 3000 1000 3000 3000 2000 3000 1000 3000 3000 0 3000 2000 3000 3000 1000 3000 500 3000 3000 2000 3000 1000 3000 0
