+++ heudiconv_app/bids-jsons/site-RU_dwi.bvec
@@ -1,0 +1,3 @@
+0 0 -0.35632 -0.93454 -0.186892 -0.110725 -0.186718 -0.86745 -0.107866 -0.586807 -0.349101 -0.587493 -0.470037 0.664209 -0.0632327 0.611197 0.84339 -0.498872 0.972027 -0.846114 -0.618066 0.654896 0.549955 0.636556 -0.676535 -0.91633 0.680791 -0.820527 0.71969 -0.25813 0.0399953 -0.257834 0.200595 -0.304395 0 0.734062 -0.343883 0.224275 -0.3446 0.560945 0.33595 -0.825864 0.765487 -0.292742 -0.807288 0.585123 0.291464 0.942971 0 -0.168609 -0.788488 -0.361622 -0.253145 -0.361407 0.00403519 0.915666 -0.354184 0.324211 0.35452 -0.799797 0.16773 0.383441 -0.203845 0.32217 -0.4324 0.415006 0.322068 -0.511952 0 0.522689 0.762158 -0.590558 0.710463 -0.759779 0.0801102 0.101005 0.996227 -0.0345736 0.991607 0.286786 0.99179 -0.072738 0.520193 0.0561379 -0.353826 0.0568478 -0.580657 -0.77126 0 0.85706 0.757088 0.324678 0.423485 0.757575 0.151971 -0.853692 -0.776659 0.937844 -0.827477 -0.0135232 0.825567 -0.274053 0
+0 0 -0.656106 -0.272129 0.956638 0.119186 0.956445 0.327291 -0.39381 -0.191002 0.905651 -0.191288 0.565182 -0.557902 -0.00590469 0.648031 -0.325055 0.436111 -0.0599513 0.326353 0.28965 -0.63864 0.00872961 0.0985049 0.0228949 -0.162681 -0.0228447 0.0948112 -0.21465 -0.595429 0.740169 -0.59451 0.849021 0.943468 0 0.216075 0.417533 0.0605957 0.417648 0.81434 -0.26657 -0.399536 -0.442562 -0.736008 0.565785 -0.717459 0.732314 -0.33117 0 -0.466887 -0.48015 0.580421 -0.961664 0.579849 -0.921366 0.393356 -0.928513 0.851282 0.929297 -0.571539 -0.536658 0.787245 0.443809 -0.860533 0.144133 0.893895 -0.859895 0.796539 0 0.398173 0.526382 0.806461 0.691637 -0.524554 0.974427 -0.885159 0.0522219 0.697988 -0.0343373 0.627541 -0.0343052 0.99388 0.739823 -0.0849795 0.703411 -0.0852574 -0.503966 -0.371767 0 0.0507452 -0.630829 0.17003 0.312857 -0.630505 -0.885934 0.153502 0.606121 -0.216076 -0.491673 0.210203 0.490492 0.786637 0
+0 0 0.665253 0.229305 -0.223417 0.986679 -0.224387 0.374714 -0.912841 0.786877 -0.240676 0.786296 -0.677963 -0.497566 -0.997981 -0.454416 -0.427823 0.748955 -0.227091 0.421407 -0.730819 0.404041 0.835149 -0.764914 -0.736054 -0.365887 0.732121 -0.56369 -0.660282 -0.76081 -0.67123 -0.761629 0.488799 0.131196 0 0.643789 0.841077 0.97264 0.840726 -0.148968 0.90337 -0.397894 0.467085 0.610405 -0.167849 0.377999 -0.615439 -0.0336537 0 0.868094 -0.384373 -0.729617 0.105447 -0.730178 0.388674 0.0826254 -0.111438 -0.41256 0.103551 0.183489 0.826961 -0.482927 0.872629 -0.394574 0.890088 0.169473 -0.396047 0.321608 0 -0.753827 0.376879 -0.0293518 0.129928 -0.38416 0.209939 -0.454193 0.0693105 0.715275 0.124649 0.723841 0.123189 -0.0831341 0.426687 0.9948 0.616457 0.994736 -0.639418 0.516669 0 -0.512712 0.169918 -0.930416 0.850166 0.16895 -0.438208 0.497642 0.171518 0.271588 0.271182 -0.977564 -0.279029 -0.553261 0
