+++ heudiconv_app/bids-jsons/site-RU_dwi.json
@@ -1,0 +1,176 @@
+{
+    "AcquisitionDuration": 453.025,
+    "AcquisitionMatrixPE": 140,
+    "AcquisitionNumber": 1,
+    "AcquisitionTime": "12:05:0.815000",
+    "BandwidthPerPixelPhaseEncode": 10.661,
+    "BaseResolution": 140,
+    "BodyPartExamined": "BRAIN",
+    "CoilCombinationMethod": "Adaptive Combine",
+    "CoilString": "HeadNeck_64",
+    "ConversionSoftware": "dcm2niix",
+    "ConversionSoftwareVersion": "v1.0.20220720",
+    "DerivedVendorReportedEchoSpacing": 0.000669999,
+    "DeviceSerialNumber": "66022",
+    "DiffusionScheme": "Monopolar",
+    "DwellTime": 2.1e-06,
+    "EchoTime": 0.089,
+    "EchoTrainLength": 105,
+    "EffectiveEchoSpacing": 0.000669999,
+    "FlipAngle": 90,
+    "ImageOrientationPatientDICOM": [
+        1,
+        0,
+        0,
+        0,
+        1,
+        0
+    ],
+    "ImageType": [
+        "ORIGINAL",
+        "PRIMARY",
+        "DIFFUSION",
+        "NONE"
+    ],
+    "ImageTypeText": [
+        "ORIGINAL",
+        "PRIMARY",
+        "DIFFUSION",
+        "NONE",
+        "ND"
+    ],
+    "ImagingFrequency": 123.257,
+    "InPlanePhaseEncodingDirectionDICOM": "COL",
+    "InstitutionAddress": "East Beltline NE 2750,Grand Rapids,MI,US,49503",
+    "InstitutionName": "Spectrum ICCB Campus",
+    "MRAcquisitionType": "2D",
+    "MTState": false,
+    "MagneticFieldStrength": 3,
+    "Manufacturer": "Siemens",
+    "ManufacturersModelName": "MAGNETOM Prisma",
+    "Modality": "MR",
+    "MultibandAccelerationFactor": 3,
+    "NonlinearGradientCorrection": false,
+    "ParallelAcquisitionTechnique": "SMS",
+    "PartialFourier": 0.75,
+    "PartialFourierDirection": "PHASE",
+    "PatientPosition": "HFS",
+    "PercentPhaseFOV": 100,
+    "PercentSampling": 100,
+    "PhaseEncodingDirection": "j-",
+    "PhaseResolution": 1,
+    "PixelBandwidth": 1701,
+    "ProcedureStepDescription": "Physicist ABCD",
+    "ProtocolName": "DWI",
+    "PulseSequenceDetails": "%SiemensSeq%\\ep2d_diff",
+    "PulseSequenceName": "*epse2d1_140",
+    "ReceiveCoilActiveElements": "HC1-7",
+    "ReceiveCoilName": "HeadNeck_64",
+    "ReconMatrixPE": 140,
+    "RefLinesPE": 17,
+    "RepetitionTime": 4.2,
+    "ScanOptions": "PFP\\FS",
+    "ScanningSequence": "EP",
+    "SequenceVariant": "SK\\SP",
+    "SeriesDescription": "DWI",
+    "SeriesNumber": 6,
+    "ShimSetting": [
+        -1205,
+        -11038,
+        2823,
+        143,
+        166,
+        -465,
+        108,
+        64
+    ],
+    "SliceThickness": 1.7,
+    "SliceTiming": [
+        0,
+        2.1675,
+        0.155,
+        2.3225,
+        0.31,
+        2.475,
+        0.4625,
+        2.63,
+        0.6175,
+        2.785,
+        0.7725,
+        2.94,
+        0.9275,
+        3.095,
+        1.0825,
+        3.25,
+        1.2375,
+        3.405,
+        1.3925,
+        3.56,
+        1.5475,
+        3.715,
+        1.7025,
+        3.87,
+        1.8575,
+        4.025,
+        2.0125,
+        0,
+        2.1675,
+        0.155,
+        2.3225,
+        0.31,
+        2.475,
+        0.4625,
+        2.63,
+        0.6175,
+        2.785,
+        0.7725,
+        2.94,
+        0.9275,
+        3.095,
+        1.0825,
+        3.25,
+        1.2375,
+        3.405,
+        1.3925,
+        3.56,
+        1.5475,
+        3.715,
+        1.7025,
+        3.87,
+        1.8575,
+        4.025,
+        2.0125,
+        0,
+        2.1675,
+        0.155,
+        2.3225,
+        0.31,
+        2.475,
+        0.4625,
+        2.63,
+        0.6175,
+        2.785,
+        0.7725,
+        2.94,
+        0.9275,
+        3.095,
+        1.0825,
+        3.25,
+        1.2375,
+        3.405,
+        1.3925,
+        3.56,
+        1.5475,
+        3.715,
+        1.7025,
+        3.87,
+        1.8575,
+        4.025,
+        2.0125
+    ],
+    "SoftwareVersions": "syngo MR XA30",
+    "SpacingBetweenSlices": 1.7,
+    "StationName": "AWP66022",
+    "TotalReadoutTime": 0.0931298,
+    "TxRefAmp": 147.376
+}
