+++ heudiconv_app/bids-jsons/site-RU_task-cuff_bold.json
@@ -1,0 +1,154 @@
+{
+    "AcquisitionDuration": 367.702,
+    "AcquisitionMatrixPE": 90,
+    "AcquisitionNumber": 1,
+    "AcquisitionTime": "12:16:22.462500",
+    "BandwidthPerPixelPhaseEncode": 21.786,
+    "BaseResolution": 90,
+    "BodyPartExamined": "BRAIN",
+    "CoilCombinationMethod": "Sum of Squares",
+    "CoilString": "HeadNeck_64",
+    "ConversionSoftware": "dcm2niix",
+    "ConversionSoftwareVersion": "v1.0.20220720",
+    "DerivedVendorReportedEchoSpacing": 0.000510012,
+    "DeviceSerialNumber": "66022",
+    "DwellTime": 2e-06,
+    "EchoTime": 0.03,
+    "EchoTrainLength": 90,
+    "EffectiveEchoSpacing": 0.000510012,
+    "FlipAngle": 52,
+    "ImageOrientationPatientDICOM": [
+        1,
+        0,
+        0,
+        0,
+        1,
+        0
+    ],
+    "ImageType": [
+        "ORIGINAL",
+        "PRIMARY",
+        "FMRI",
+        "NONE"
+    ],
+    "ImageTypeText": [
+        "ORIGINAL",
+        "PRIMARY",
+        "M",
+        "ND"
+    ],
+    "ImagingFrequency": 123.257,
+    "InPlanePhaseEncodingDirectionDICOM": "COL",
+    "InstitutionAddress": "East Beltline NE 2750,Grand Rapids,MI,US,49503",
+    "InstitutionName": "Spectrum ICCB Campus",
+    "MRAcquisitionType": "2D",
+    "MTState": false,
+    "MagneticFieldStrength": 3,
+    "Manufacturer": "Siemens",
+    "ManufacturersModelName": "MAGNETOM Prisma",
+    "Modality": "MR",
+    "MultibandAccelerationFactor": 6,
+    "NonlinearGradientCorrection": false,
+    "ParallelAcquisitionTechnique": "SMS",
+    "PartialFourier": 1,
+    "PatientPosition": "HFS",
+    "PercentPhaseFOV": 100,
+    "PercentSampling": 100,
+    "PhaseEncodingDirection": "j-",
+    "PhaseResolution": 1,
+    "PixelBandwidth": 2778,
+    "ProcedureStepDescription": "Physicist ABCD",
+    "ProtocolName": "CUFF1",
+    "PulseSequenceDetails": "%SiemensSeq%\\ep2d_bold",
+    "PulseSequenceName": "*epfid2d1_90",
+    "ReceiveCoilActiveElements": "HC1-7",
+    "ReceiveCoilName": "HeadNeck_64",
+    "ReconMatrixPE": 90,
+    "RefLinesPE": 22,
+    "RepetitionTime": 0.8,
+    "ScanOptions": "FS",
+    "ScanningSequence": "EP",
+    "SequenceVariant": "SK",
+    "SeriesDescription": "CUFF1",
+    "SeriesNumber": 15,
+    "ShimSetting": [
+        -1205,
+        -11038,
+        2823,
+        141,
+        166,
+        -466,
+        108,
+        64
+    ],
+    "SliceThickness": 2.4,
+    "SliceTiming": [
+        0.5475,
+        0,
+        0.39,
+        0.0775005,
+        0.4675,
+        0.235,
+        0.625,
+        0.3125,
+        0.7025,
+        0.1575,
+        0.5475,
+        0,
+        0.39,
+        0.0775005,
+        0.4675,
+        0.235,
+        0.625,
+        0.3125,
+        0.7025,
+        0.1575,
+        0.5475,
+        0,
+        0.39,
+        0.0775005,
+        0.4675,
+        0.235,
+        0.625,
+        0.3125,
+        0.7025,
+        0.1575,
+        0.5475,
+        0,
+        0.39,
+        0.0775005,
+        0.4675,
+        0.235,
+        0.625,
+        0.3125,
+        0.7025,
+        0.1575,
+        0.5475,
+        0,
+        0.39,
+        0.0775005,
+        0.4675,
+        0.235,
+        0.625,
+        0.3125,
+        0.7025,
+        0.1575,
+        0.5475,
+        0,
+        0.39,
+        0.0775005,
+        0.4675,
+        0.235,
+        0.625,
+        0.3125,
+        0.7025,
+        0.1575
+    ],
+    "SoftwareVersions": "syngo MR XA30",
+    "SpacingBetweenSlices": 2.4,
+    "StationName": "AWP66022",
+    "TaskName": "cuff",
+    "TotalReadoutTime": 0.045391,
+    "TxRefAmp": 147.376,
+    "VendorReportedEchoSpacing": 0.00051
+}
