+++ heudiconv_app/bids-jsons/site-RUphantom_T1w.json
@@ -1,0 +1,84 @@
+{
+    "AcquisitionDuration": 62.382,
+    "AcquisitionMatrixPE": 208,
+    "AcquisitionNumber": 1,
+    "AcquisitionTime": "10:40:41.835000",
+    "BaseResolution": 256,
+    "BodyPartExamined": "HEAD",
+    "CoilCombinationMethod": "Sum of Squares",
+    "CoilString": "HeadNeck_64",
+    "ConversionSoftware": "dcm2niix",
+    "ConversionSoftwareVersion": "v1.0.20220720",
+    "DeviceSerialNumber": "66022",
+    "DwellTime": 8.2e-06,
+    "EchoTime": 0.00614,
+    "FlipAngle": 15,
+    "ImageOrientationPatientDICOM": [
+        1,
+        0,
+        0,
+        0,
+        1,
+        0
+    ],
+    "ImageType": [
+        "ORIGINAL",
+        "PRIMARY",
+        "M",
+        "NONE"
+    ],
+    "ImageTypeText": [
+        "ORIGINAL",
+        "PRIMARY",
+        "M",
+        "NORM",
+        "DIS2D"
+    ],
+    "ImagingFrequency": 123.257,
+    "InPlanePhaseEncodingDirectionDICOM": "COL",
+    "InstitutionAddress": "East Beltline NE 2750,Grand Rapids,MI,US,49503",
+    "InstitutionName": "Spectrum ICCB Campus",
+    "MRAcquisitionType": "2D",
+    "MTState": false,
+    "MagneticFieldStrength": 3,
+    "Manufacturer": "Siemens",
+    "ManufacturersModelName": "MAGNETOM Prisma",
+    "MatrixCoilMode": "SENSE",
+    "Modality": "MR",
+    "NonlinearGradientCorrection": true,
+    "NumberOfAverages": 2,
+    "PartialFourier": 1,
+    "PatientPosition": "HFS",
+    "PercentPhaseFOV": 81.25,
+    "PercentSampling": 100,
+    "PhaseEncodingDirection": "j-",
+    "PhaseResolution": 1,
+    "PixelBandwidth": 238,
+    "ProcedureStepDescription": "Brain Spectrum non DOT Brain",
+    "ProtocolName": "anat-T1w_acq_GRE",
+    "PulseSequenceDetails": "%SiemensSeq%\\gre",
+    "PulseSequenceName": "*fl2d1r",
+    "ReceiveCoilActiveElements": "HC1-7",
+    "ReceiveCoilName": "HeadNeck_64",
+    "ReconMatrixPE": 208,
+    "RepetitionTime": 0.15,
+    "ScanningSequence": "GR",
+    "SequenceVariant": "SP",
+    "SeriesDescription": "anat-T1w_acq_GRE",
+    "SeriesNumber": 2,
+    "ShimSetting": [
+        -1200,
+        -11047,
+        2840,
+        152,
+        177,
+        -293,
+        146,
+        43
+    ],
+    "SliceThickness": 5,
+    "SoftwareVersions": "syngo MR XA30",
+    "SpacingBetweenSlices": 5,
+    "StationName": "AWP66022",
+    "TxRefAmp": 150.636
+}
