+++ heudiconv_app/bids-jsons/site-RUphantom_acq-b1000_dwi.bval
@@ -1,0 +1,1 @@
+0 1000 1000 1000 1000 1000 1000 0 1000 1000 1000 1000 1000 1000
