+++ heudiconv_app/bids-jsons/site-RUphantom_acq-b1000_dwi.bvec
@@ -1,0 +1,3 @@
+0 0.707462 -0.707462 0.000705688 0.000706392 0.706045 -0.706045 0 0.707462 -0.707462 0.000705688 0.000706392 0.706045 -0.706045
+0 -0.00141997 -0.00141997 0.708519 0.708522 0.708162 0.708162 0 -0.00141997 -0.00141997 0.708519 0.708522 0.708162 0.708162
+0 -0.70675 -0.70675 -0.705691 0.705689 -0.00282415 -0.00282415 0 -0.70675 -0.70675 -0.705691 0.705689 -0.00282415 -0.00282415
