+++ heudiconv_app/bids-jsons/site-RUphantom_acq-b1000_dwi.json
@@ -1,0 +1,97 @@
+{
+    "AcquisitionDuration": 35.32,
+    "AcquisitionMatrixPE": 130,
+    "AcquisitionNumber": 1,
+    "AcquisitionTime": "10:42:10.737500",
+    "BandwidthPerPixelPhaseEncode": 10.395,
+    "BaseResolution": 130,
+    "BodyPartExamined": "HEAD",
+    "CoilCombinationMethod": "Adaptive Combine",
+    "CoilString": "HeadNeck_64",
+    "ConversionSoftware": "dcm2niix",
+    "ConversionSoftwareVersion": "v1.0.20220720",
+    "DerivedVendorReportedEchoSpacing": 0.000740001,
+    "DeviceSerialNumber": "66022",
+    "DiffusionScheme": "Bipolar",
+    "DwellTime": 2.6e-06,
+    "EchoTime": 0.091,
+    "EchoTrainLength": 98,
+    "EffectiveEchoSpacing": 0.000740001,
+    "FlipAngle": 90,
+    "ImageOrientationPatientDICOM": [
+        1,
+        0,
+        0,
+        0,
+        1,
+        0
+    ],
+    "ImageType": [
+        "ORIGINAL",
+        "PRIMARY",
+        "DIFFUSION",
+        "NONE"
+    ],
+    "ImageTypeText": [
+        "ORIGINAL",
+        "PRIMARY",
+        "DIFFUSION",
+        "NONE",
+        "ND",
+        "NORM"
+    ],
+    "ImagingFrequency": 123.257,
+    "InPlanePhaseEncodingDirectionDICOM": "COL",
+    "InstitutionAddress": "East Beltline NE 2750,Grand Rapids,MI,US,49503",
+    "InstitutionName": "Spectrum ICCB Campus",
+    "MRAcquisitionType": "2D",
+    "MTState": false,
+    "MagneticFieldStrength": 3,
+    "Manufacturer": "Siemens",
+    "ManufacturersModelName": "MAGNETOM Prisma",
+    "MatrixCoilMode": "SENSE",
+    "Modality": "MR",
+    "NonlinearGradientCorrection": false,
+    "PartialFourier": 0.75,
+    "PartialFourierDirection": "PHASE",
+    "PatientPosition": "HFS",
+    "PercentPhaseFOV": 100,
+    "PercentSampling": 100,
+    "PhaseEncodingDirection": "j-",
+    "PhaseResolution": 1,
+    "PixelBandwidth": 1479,
+    "ProcedureStepDescription": "Brain Spectrum non DOT Brain",
+    "ProtocolName": "dwi_dwi_acq_b1000",
+    "PulseSequenceDetails": "%SiemensSeq%\\ep2d_diff",
+    "PulseSequenceName": "*epse2d1_130",
+    "ReceiveCoilActiveElements": "HC1-7",
+    "ReceiveCoilName": "HeadNeck_64",
+    "ReconMatrixPE": 130,
+    "RepetitionTime": 2,
+    "ScanOptions": "PFP\\FS",
+    "ScanningSequence": "EP",
+    "SequenceVariant": "SK\\SP",
+    "SeriesDescription": "dwi_dwi_acq_b1000",
+    "SeriesNumber": 3,
+    "ShimSetting": [
+        -1200,
+        -11038,
+        2833,
+        129,
+        96,
+        -547,
+        62,
+        49
+    ],
+    "SliceThickness": 5,
+    "SliceTiming": [
+        0,
+        1.32,
+        0.66
+    ],
+    "SoftwareVersions": "syngo MR XA30",
+    "SpacingBetweenSlices": 6,
+    "StationName": "AWP66022",
+    "TotalReadoutTime": 0.0954601,
+    "TxRefAmp": 150.636
+}
