+++ heudiconv_app/bids-jsons/site-RUphantom_acq-b2000_dwi.bval
@@ -1,0 +1,1 @@
+0 2000 2000 2000 2000 2000 2000 0 2000 2000 2000 2000 2000 2000
