+++ heudiconv_app/bids-jsons/site-RUphantom_acq-b2000_dwi.bvec
@@ -1,0 +1,3 @@
+0 0.707461 -0.707461 0.000353022 0.000353022 0.7064 -0.7064 0 0.707461 -0.707461 0.000353022 0.000353022 0.7064 -0.7064
+0 -0.00106294 -0.00106294 0.708167 0.708167 0.707812 0.707812 0 -0.00106294 -0.00106294 0.708167 0.708167 0.707812 0.707812
+0 -0.706752 -0.706752 -0.706045 0.706045 -0.00141209 -0.00141209 0 -0.706752 -0.706752 -0.706045 0.706045 -0.00141209 -0.00141209
