+++ heudiconv_app/bids-jsons/site-RUphantom_task-rest_bold.json
@@ -1,0 +1,148 @@
+{
+    "AcquisitionDuration": 367.692,
+    "AcquisitionMatrixPE": 90,
+    "AcquisitionNumber": 1,
+    "AcquisitionTime": "10:43:27.440000",
+    "BandwidthPerPixelPhaseEncode": 20.576,
+    "BaseResolution": 90,
+    "BodyPartExamined": "HEAD",
+    "CoilCombinationMethod": "Sum of Squares",
+    "CoilString": "HeadNeck_64",
+    "ConversionSoftware": "dcm2niix",
+    "ConversionSoftwareVersion": "v1.0.20220720",
+    "DerivedVendorReportedEchoSpacing": 0.000540003,
+    "DeviceSerialNumber": "66022",
+    "DwellTime": 2.2e-06,
+    "EchoTime": 0.032,
+    "EchoTrainLength": 90,
+    "EffectiveEchoSpacing": 0.000540003,
+    "FlipAngle": 52,
+    "ImageOrientationPatientDICOM": [
+        1,
+        0,
+        0,
+        0,
+        1,
+        0
+    ],
+    "ImageType": [
+        "ORIGINAL",
+        "PRIMARY",
+        "FMRI",
+        "NONE"
+    ],
+    "ImageTypeText": [
+        "ORIGINAL",
+        "PRIMARY",
+        "M",
+        "ND"
+    ],
+    "ImagingFrequency": 123.257,
+    "InPlanePhaseEncodingDirectionDICOM": "COL",
+    "InstitutionAddress": "East Beltline NE 2750,Grand Rapids,MI,US,49503",
+    "InstitutionName": "Spectrum ICCB Campus",
+    "MRAcquisitionType": "2D",
+    "MTState": false,
+    "MagneticFieldStrength": 3,
+    "Manufacturer": "Siemens",
+    "ManufacturersModelName": "MAGNETOM Prisma",
+    "Modality": "MR",
+    "MultibandAccelerationFactor": 6,
+    "NonlinearGradientCorrection": false,
+    "ParallelAcquisitionTechnique": "SMS",
+    "PartialFourier": 1,
+    "PatientPosition": "HFS",
+    "PercentPhaseFOV": 100,
+    "PercentSampling": 100,
+    "PhaseEncodingDirection": "j",
+    "PhaseResolution": 1,
+    "PixelBandwidth": 2525,
+    "ProcedureStepDescription": "Brain Spectrum non DOT Brain",
+    "ProtocolName": "func_bold_acq_QA",
+    "PulseSequenceDetails": "%SiemensSeq%\\ep2d_bold",
+    "PulseSequenceName": "*epfid2d1_90",
+    "ReceiveCoilActiveElements": "HC1-7",
+    "ReceiveCoilName": "HeadNeck_64",
+    "ReconMatrixPE": 90,
+    "RefLinesPE": 22,
+    "RepetitionTime": 0.8,
+    "ScanOptions": "FS",
+    "ScanningSequence": "EP",
+    "SequenceVariant": "SK",
+    "SeriesDescription": "func_bold_acq_QA",
+    "SeriesNumber": 9,
+    "ShimSetting": [
+        -1198,
+        -11041,
+        2822,
+        135,
+        160,
+        -623,
+        84,
+        56
+    ],
+    "SliceThickness": 2.4,
+    "SliceTiming": [
+        0,
+        0.435,
+        0.0875,
+        0.52,
+        0.175,
+        0.6075,
+        0.26,
+        0.695,
+        0.3475,
+        0,
+        0.435,
+        0.0875,
+        0.52,
+        0.175,
+        0.6075,
+        0.26,
+        0.695,
+        0.3475,
+        0,
+        0.435,
+        0.0875,
+        0.52,
+        0.175,
+        0.6075,
+        0.26,
+        0.695,
+        0.3475,
+        0,
+        0.435,
+        0.0875,
+        0.52,
+        0.175,
+        0.6075,
+        0.26,
+        0.695,
+        0.3475,
+        0,
+        0.435,
+        0.0875,
+        0.52,
+        0.175,
+        0.6075,
+        0.26,
+        0.695,
+        0.3475,
+        0,
+        0.435,
+        0.0875,
+        0.52,
+        0.175,
+        0.6075,
+        0.26,
+        0.695,
+        0.3475
+    ],
+    "SoftwareVersions": "syngo MR XA30",
+    "SpacingBetweenSlices": 2.4,
+    "StationName": "AWP66022",
+    "TaskName": "rest",
+    "TotalReadoutTime": 0.0480603,
+    "TxRefAmp": 150.636,
+    "VendorReportedEchoSpacing": 0.00054
+}
