+++ heudiconv_app/env-v1.0.20211006.yml
@@ -1,0 +1,30 @@
+name: v1.0.20211006
+channels:
+  - conda-forge
+dependencies:
+  - python
+  - traits>=4.6
+  - scipy
+  - numpy
+  - pandas
+  - nomkl
+  - nilearn
+  - pydicom
+  - openpyxl 
+  - deepdiff
+  - pigz
+  - git-annex
+  - nipype
+  - heudiconv=0.10.0
+  - dcmstack
+  - sqlalchemy<1.4.0
+  - cmake
+  - compilers
+  - make
+  # pybids suff
+  - docopt
+  - num2words
+  - wrapt>=1.0
+  - astor>=0.8
+  - formulaic<0.4
+  - nodejs # bids-validator
