+++ heudiconv_app/env.yml
@@ -1,0 +1,31 @@
+name: v1.0.20220720
+channels:
+  - conda-forge
+dependencies:
+  - python
+  - traits>=4.6
+  - scipy
+  - numpy
+  - pandas
+  - nomkl
+  - nilearn
+  - pydicom
+  - openpyxl 
+  - deepdiff
+  - pigz
+  - dcm2niix=1.0.20220720
+  - git-annex
+  - nipype
+  - heudiconv=0.10.0 
+  - dcmstack
+  - sqlalchemy<1.4.0
+  - cmake
+  - compilers
+  - make
+  # pybids suff
+  - docopt
+  - num2words
+  - wrapt>=1.0
+  - astor>=0.8
+  - formulaic<0.4
+  - nodejs # bids-validator
